change type check to int64_t

+++ src/graph/transform/kahip_partition_hetero.cc
@@ -34,7 +34,7 @@ IdArray KaHIPPartition(
   int64_t * part = static_cast<int64_t *>(part_arr->data);
   double imbalance = 0.03;
   int64_t vwgt_len = vwgt_arr->shape[0];
-  CHECK_EQ(sizeof(int), vwgt_arr->dtype.bits / 8)
+  CHECK_EQ(sizeof(int64_t), vwgt_arr->dtype.bits / 8)
       << "The vertex weight array doesn't have right type";
   CHECK(vwgt_len % g->NumVertices(0) == 0)
       << "The vertex weight array doesn't have right number of elements";
